requirements.txt
PyYAML==6.0.1
PySide6==6.9.1
